created folder for SauceDemo and moved tests structure
