+++ cypress/e2e/SauceDemo/UI/PageObjects/InventoryPage.js
@@ -1,5 +1,5 @@
 /// <reference types="cypress" />
-import settings from '../../../fixtures/settings.json'
+import settings from '../../../../fixtures/settings.json'
 
 class InventoryPage {
 
@@ -8,7 +8,6 @@ class InventoryPage {
     }
 
     TitleField() {return cy.get('span[class="title"]');}
-
 
 }
 module.exports = new InventoryPage();
